# Backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Alfa HAK Copilot API
  description: |
    Swagger спецификация для продакшен-инстанса Alfa HAK Copilot.

    Бэкенд основан на FastAPI и обслуживает Telegram-бота, мини-веб-приложение,
    загрузку документов, калькулятор и интеграцию с Альфа-Бизнесом.
  version: "1.0.0"
  contact:
    name: Alfa HAK Copilot Team
    url: https://alpha-hak-copilot.ruka.me
servers:
  - url: https://api-alpha-hak-copilot.ruka.me/api
    description: Production (24/7)
  - url: http://localhost:8000/api
    description: Local development
...
paths:
...
  /knowledge/documents:
    post:
      tags: [knowledge]
      summary: Загрузка и индексация документа
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [user_id, file, title]
              properties:
                user_id:
                  type: string
                  description: Telegram user id
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                category:
                  type: string
                  default: general
                tags_json:
                  type: string
                  description: JSON-массив тегов
                  example: '["finance", "risk"]'
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Документ принят
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSource'
    get:
      tags: [knowledge]
      summary: Список документов
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
          required: false
          description: Фильтр по владельцу
      responses:
        '200':
          description: Коллекция документов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DocumentSource'
  /knowledge/search:
    get:
      tags: [knowledge]
      summary: Векторный поиск по памяти
      parameters:
        - in: query
          name: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Результаты поиска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeSearchResponse'
  /chat/messages:
    post:
      tags: [chat]
      summary: Отправить сообщение ИИ-оркестратору
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Ответ ассистента или план калькулятора
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
  /chat/execute:
    post:
      tags: [chat]
      summary: Выполнить подтверждённый калькуляторный план
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculatorExecutionRequest'
      responses:
        '200':
          description: Результат расчёта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '404':
          description: План не найден или истёк
  /chat/context/{user_id}:
    delete:
      tags: [chat]
      summary: Сброс контекста и диалога пользователя
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: Контекст очищен
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
        '500':
          description: Ошибка при очистке
...
components:
  parameters:
    UserIdPath:
      in: path
      name: user_id
      required: true
      schema:
        type: string
      description: Telegram user id
  schemas:
    DocumentSource:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          example: general
        owner_id:
          type: string
          nullable: true
        uploaded_at:
          type: string
          format: date-time
        size_bytes:
          type: integer
        content_type:
          type: string
        original_filename:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        checksum:
          type: string
          nullable: true
        status:
          type: string
          example: indexed
    KnowledgeSearchHit:
      type: object
      properties:
        id:
          type: string
        score:
          type: number
          format: float
        text:
          type: string
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    KnowledgeSearchResponse:
      type: object
      properties:
        query:
          type: string
        hits:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeSearchHit'
        embedding_available:
          type: boolean
          default: true
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    OrchestrationDecision:
      type: object
      properties:
        mode:
          type: string
          enum: [advisor, calculator]
        summary:
          type: string
          nullable: true
        calculator_instructions:
          type: string
          nullable: true
        tool_calls:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    ToolExecutionResult:
      type: object
      properties:
        name:
          type: string
        output:
          type: string
        success:
          type: boolean
        error:
          type: string
          nullable: true
        duration_ms:
          type: integer
          nullable: true
    ChatRequest:
      type: object
      required: [user_id, content]
      properties:
        user_id:
          type: string
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
          nullable: true
    ChatResponse:
      type: object
      properties:
        reply:
          $ref: '#/components/schemas/ChatMessage'
        decision:
          $ref: '#/components/schemas/OrchestrationDecision'
        knowledge_hits:
          type: array
          items:
            $ref: '#/components/schemas/KnowledgeSearchHit'
        tool_results:
          type: array
          items:
            $ref: '#/components/schemas/ToolExecutionResult'
    CalculatorExecutionRequest:
      type: object
      required: [plan_id, user_id]
      properties:
        plan_id:
          type: string
        user_id:
          type: string
...